Bug fix for cd'ing into directories with spaces.
The argument to cd needs to be surrouned by double quotes to enusre that the argument is treated as a single argument. This is relevant when an escape character is used to specify a directory with a space in the name. The escape character is processed when it is handed to the internal cd function. When cd was called before this commit, it would see multiple arguments because the space characters were no longer escaped.