Move editor to next or previous group hotkeys
I wanted to set this to ctrl+) for move the editor to the left and ctrl+7 for moving the editor to the right. This would line it up with the movement I'm used to. Unfortunately, there is something weird occuring when I try to use ctrl+). VSCode detects shift instead. I think this is because ) is the shifted character from 0 on a regular keyboard and the Corne's zmk firmware is doing something special to handle sending ) to the machine. I don't know exactly what's wrong though.

The next best thing was to use ctrl+5 and ctrl+6 to move the files to the next or previous group respectively. Since VSCode displays the groups vertically, this still visually makes sense, it's just not very prominent.