Setup symlinks in directories and git number
Before this, setup only created symlinks in ~/ for the files contained in home/ (this repo's home/ directory). In addition to this, it now creates directories contained in home/ that are missing in ~/ and creates symlinks for the files contained within those directories.

It also creates the necessary symlinks needed to use the git-number scripts. This is supplemented by the additional directory added to PATH.