Initial Neovim Config
This established my initiali neovim config. I've been piecing it together over the last couple of weeks without committing anyting because of how in flux and experimental all of the changes are. I've arrived at something that feels relatively complete at this point, but more changes are bound to occur.