Clear Tiling, Maximized, and Floating Controls
The way to get into tiling mode was to disable flying or maximized mode, thus the key would be dependent on which mode the client was in. Thinking about it this way was cumbersome and the code to do so was more complicated than it needed to be. Since I don't ever use minimizing intentionally, I've opted to get rid of it and make it the key to activate tiling mode on the client. This addition and the surrounding simplifications makes the operation very clear.